pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "alphasift"
version = "0.1.0"
description = "自动选股 Skill — 从全市场中按策略筛选、评分、排序候选股票"
readme = "README.md"
license = "Apache-2.0"
license-files = ["LICENSE"]
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0",
    "pyyaml>=6.0",
    "litellm>=1.0",
    "efinance>=0.4",
    "akshare>=1.10",
    "requests>=2.28",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[project.scripts]
alphasift = "alphasift.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["alphasift*"]

[tool.setuptools.package-data]
alphasift = ["strategies/*.yaml"]

[tool.setuptools.data-files]
"." = ["SKILL.md", ".env.example", "LICENSE"]
"agents" = ["agents/openai.yaml"]
